Changement de l'étape 1
Pour résoudre les dépendances on doit changer reada_elf_header de l'étape 1:
•Toutes les fonctions de lecture du Header sont dans un module readHeader.
•Ainsi, read_elf_header n'est composé que d'un main().
•Le Mafefile a été adapté en conséquence.

Ainsi, on peut utiliser la lecture du Header pour lire la table des Sections sans avoir un conflit avec deux main() comme au commit précédent.

L'étape 2 sera aussi sous forme d'un module.
